refactor(video): video repository, entity
